# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TERA EXAM Sarkari exam platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List published results
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Result" }
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get a published result
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Result" }
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Public: Admit Cards
  # ---------------------------------------------------------------------
...
  /topics/{id}/questions:
    get:
      operationId: listQuestionsByTopic
      tags: [exam-prep]
      summary: List practice questions for a topic (answers withheld)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Practice questions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PracticeQuestion" }
  /topics/{id}/submit:
    post:
      operationId: submitTopicAnswers
      tags: [exam-prep]
      summary: Submit answers for a topic practice session and get scored results
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswersInput"
      responses:
        "200":
          description: Scored result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitAnswersResult"

  # ---------------------------------------------------------------------
  # Admin: Jobs
  # ---------------------------------------------------------------------
...
  /admin/results:
    get:
      operationId: adminListResults
      tags: [admin]
      summary: List all results (any status)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PublishStatus" }
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Result" }
    post:
      operationId: createResult
      tags: [admin]
      summary: Create a result link
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResultInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Result" }
  /admin/results/{id}:
    patch:
      operationId: updateResult
      tags: [admin]
      summary: Update a result (also used for one-click publish via status field)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResultUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Result" }
        "404":
          description: Not found
    delete:
      operationId: deleteResult
      tags: [admin]
      summary: Delete a result
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Admin: Admit Cards
  # ---------------------------------------------------------------------
...
  /admin/questions:
    get:
      operationId: adminListQuestions
      tags: [admin]
      summary: List all questions (with correct answers), optionally filtered by topic
      security: [{ bearerAuth: [] }]
      parameters:
        - name: topicId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Question" }
    post:
      operationId: createQuestion
      tags: [admin]
      summary: Create a question
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuestionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
  /admin/questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [admin]
      summary: Update a question
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuestionUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [admin]
      summary: Delete a question
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Admin: Scraped item staging (approve/reject flow)
  # ---------------------------------------------------------------------
  /admin/scraped-items:
    get:
      operationId: adminListScrapedItems
      tags: [admin]
      summary: List staged scraped items
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ScrapedStatus" }
      responses:
        "200":
          description: Scraped items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScrapedItem" }
    post:
      operationId: createScrapedItem
      tags: [admin]
      summary: Manually add a staged item
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScrapedItemInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapedItem" }
  /admin/scraped-items/run-scraper:
    post:
      operationId: runScraper
      tags: [admin]
      summary: Trigger the scraper service and stage any newly found items
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Newly staged items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScrapedItem" }
  /admin/scraped-items/{id}/approve:
    patch:
      operationId: approveScrapedItem
      tags: [admin]
      summary: Approve a staged item — publishes it into the matching live table
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapedItem" }
        "404":
          description: Not found
  /admin/scraped-items/{id}/reject:
    patch:
      operationId: rejectScrapedItem
      tags: [admin]
      summary: Reject a staged item
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapedItem" }
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # NOTE: File uploads (POST /api/admin/uploads) are implemented directly in
  # the Express server with multer + Cloudinary and are intentionally NOT
  # part of this OpenAPI contract — multipart/form-data binary bodies don't
  # round-trip cleanly through the Zod/react-query codegen in this repo. The
  # frontend calls that endpoint with a plain fetch() + FormData instead of a
  # generated hook. See artifacts/api-server/src/routes/uploads.ts.
  # ---------------------------------------------------------------------
  # Admin: dashboard aggregate
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublishStatus:
      type: string
      enum: [draft, published]
    ScrapedStatus:
      type: string
      enum: [pending, approved, rejected]
    ScrapedSuggestedType:
      type: string
      enum: [job, result, admit_card]

    # --- Auth -------------------------------------------------------
...
    Result:
      type: object
      required:
        [id, title, organization, examName, resultDate, status, slug, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        organization: { type: string }
        examName: { type: string }
        resultDate: { type: string, format: date }
        resultLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ResultInput:
      type: object
      required: [title, organization, examName, resultDate, slug]
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        examName: { type: string, minLength: 1 }
        resultDate: { type: string, format: date }
        resultLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string, minLength: 1 }
    ResultUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        examName: { type: string, minLength: 1 }
        resultDate: { type: string, format: date }
        resultLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string, minLength: 1 }

    # --- AdmitCard --------------------------------------------------
...
    Question:
      type: object
      required:
        [id, topicId, questionText, options, correctOptionIndex, difficulty, createdAt]
      properties:
        id: { type: integer }
        topicId: { type: integer }
        questionText: { type: string }
        options:
          type: array
          items: { type: string }
        correctOptionIndex: { type: integer }
        explanation: { type: ["string", "null"] }
        difficulty: { type: string }
        createdAt: { type: string, format: date-time }
    QuestionInput:
      type: object
      required: [topicId, questionText, options, correctOptionIndex, difficulty]
      properties:
        topicId: { type: integer }
        questionText: { type: string, minLength: 1 }
        options:
          type: array
          items: { type: string }
          minItems: 2
        correctOptionIndex: { type: integer, minimum: 0 }
        explanation: { type: ["string", "null"] }
        difficulty: { type: string, minLength: 1 }
    QuestionUpdate:
      type: object
      properties:
        topicId: { type: integer }
        questionText: { type: string, minLength: 1 }
        options:
          type: array
          items: { type: string }
          minItems: 2
        correctOptionIndex: { type: integer, minimum: 0 }
        explanation: { type: ["string", "null"] }
        difficulty: { type: string, minLength: 1 }
    PracticeQuestion:
      type: object
      required: [id, topicId, questionText, options, difficulty]
      properties:
        id: { type: integer }
        topicId: { type: integer }
        questionText: { type: string }
        options:
          type: array
          items: { type: string }
        difficulty: { type: string }
    SubmitAnswersInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
    AnswerInput:
      type: object
      required: [questionId, selectedOption]
      properties:
        questionId: { type: integer }
        selectedOption: { type: integer }
    SubmitAnswersResult:
      type: object
      required: [score, total, results]
      properties:
        score: { type: integer }
        total: { type: integer }
        results:
          type: array
          items:
            $ref: "#/components/schemas/AnswerResult"
    AnswerResult:
      type: object
      required: [questionId, correct, selectedOption, correctOption]
      properties:
        questionId: { type: integer }
        correct: { type: boolean }
        selectedOption: { type: integer }
        correctOption: { type: integer }
        explanation: { type: ["string", "null"] }

    # --- Scraped items --------------------------------------------------
    ScrapedItem:
      type: object
      required:
        [id, sourceUrl, suggestedType, rawTitle, status, scrapedAt]
      properties:
        id: { type: integer }
        sourceUrl: { type: string }
        suggestedType: { $ref: "#/components/schemas/ScrapedSuggestedType" }
        rawTitle: { type: string }
        rawContent: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/ScrapedStatus" }
        scrapedAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
    ScrapedItemInput:
      type: object
      required: [sourceUrl, suggestedType, rawTitle]
      properties:
        sourceUrl: { type: string, minLength: 1 }
        suggestedType: { $ref: "#/components/schemas/ScrapedSuggestedType" }
        rawTitle: { type: string, minLength: 1 }
        rawContent: { type: ["string", "null"] }

    # --- Uploads --------------------------------------------------------
...